refactor: svg 아이콘 import 경로 수정
